requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Google Cloud
google-cloud-firestore==2.14.0
google-cloud-secret-manager==2.18.0
google-cloud-storage==2.14.0

# Data Processing
pandas==2.1.4
numpy==1.26.3

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2

# Logging
structlog==24.1.0

# Testing (optional)
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
